@@ -18,6 +18,9 @@ const (
 	streamZoneSync = "zone_test_sync"
 	locationZone   = "location_test"
 	resolverMetric = "resolver_test"
+	reqZone        = "one"
+	connZone       = "addr"
+	streamConnZone = "addr_stream"
 )
 
 var (
@@ -69,7 +72,7 @@ func TestStreamClient(t *testing.T) {
 
 	streamServers, err := c.GetStreamServers(streamUpstream)
 	if err != nil {
-		t.Errorf("Error getting stream servers: %w", err)
+		t.Errorf("Error getting stream servers: %v", err)
 	}
 	if len(streamServers) != 0 {
 		t.Errorf("Expected 0 servers, got %v", streamServers)
@@ -275,7 +278,7 @@ func TestStreamUpstreamServer(t *testing.T) {
 	}
 	err = c.AddStreamServer(streamUpstream, streamServer)
 	if err != nil {
-		t.Errorf("Error adding upstream server: %w", err)
+		t.Errorf("Error adding upstream server: %v", err)
 	}
 	servers, err := c.GetStreamServers(streamUpstream)
 	if err != nil {
@@ -294,7 +297,7 @@ func TestStreamUpstreamServer(t *testing.T) {
 	// remove stream upstream servers
 	_, _, _, err = c.UpdateStreamServers(streamUpstream, []client.StreamUpstreamServer{})
 	if err != nil {
-		t.Errorf("Couldn't remove servers: %w", err)
+		t.Errorf("Couldn't remove servers: %v", err)
 	}
 }
 
@@ -551,7 +554,7 @@ func TestUpstreamServer(t *testing.T) {
 	}
 	err = c.AddHTTPServer(upstream, server)
 	if err != nil {
-		t.Errorf("Error adding upstream server: %w", err)
+		t.Errorf("Error adding upstream server: %v", err)
 	}
 	servers, err := c.GetHTTPServers(upstream)
 	if err != nil {
@@ -570,7 +573,7 @@ func TestUpstreamServer(t *testing.T) {
 	// remove upstream servers
 	_, _, _, err = c.UpdateHTTPServers(upstream, []client.UpstreamServer{})
 	if err != nil {
-		t.Errorf("Couldn't remove servers: %w", err)
+		t.Errorf("Couldn't remove servers: %v", err)
 	}
 }
 
@@ -586,12 +589,12 @@ func TestStats(t *testing.T) {
 	}
 	err = c.AddHTTPServer(upstream, server)
 	if err != nil {
-		t.Errorf("Error adding upstream server: %w", err)
+		t.Errorf("Error adding upstream server: %v", err)
 	}
 
 	stats, err := c.GetStats()
 	if err != nil {
-		t.Errorf("Error getting stats: %w", err)
+		t.Errorf("Error getting stats: %v", err)
 	}
 
 	// NginxInfo
@@ -654,10 +657,13 @@ func TestStats(t *testing.T) {
 		if val.Requests < 1 {
 			t.Errorf("ServerZone stats missing: %v", val)
 		}
+		if val.Responses.Codes.HTTPOk < 1 {
+			t.Errorf("ServerZone response codes missing: %v", val.Responses.Codes)
+		}
 	} else {
 		t.Errorf("ServerZone 'test' not found")
 	}
-	if ups, ok := stats.Upstreams["test"]; ok {
+	if ups, ok := stats.Upstreams[upstream]; ok {
 		if len(ups.Peers) < 1 {
 			t.Errorf("upstream server not visible in stats")
 		} else {
@@ -686,10 +692,26 @@ func TestStats(t *testing.T) {
 		t.Errorf("Resolver %v not found", resolverMetric)
 	}
 
+	if reqLimit, ok := stats.HTTPLimitRequests[reqZone]; ok {
+		if reqLimit.Passed < 1 {
+			t.Errorf("HTTP Reqs limit stats missing: %v", reqLimit.Passed)
+		}
+	} else {
+		t.Errorf("HTTP Reqs limit %v not found", reqLimit)
+	}
+
+	if connLimit, ok := stats.HTTPLimitConnections[connZone]; ok {
+		if connLimit.Passed < 1 {
+			t.Errorf("HTTP Limit connections stats missing: %v", connLimit.Passed)
+		}
+	} else {
+		t.Errorf("HTTP Limit connections %v not found", connLimit)
+	}
+
 	// cleanup upstream servers
 	_, _, _, err = c.UpdateHTTPServers(upstream, []client.UpstreamServer{})
 	if err != nil {
-		t.Errorf("Couldn't remove servers: %w", err)
+		t.Errorf("Couldn't remove servers: %v", err)
 	}
 }
 
@@ -720,7 +742,7 @@ func TestUpstreamServerDefaultParameters(t *testing.T) {
 	}
 	err = c.AddHTTPServer(upstream, server)
 	if err != nil {
-		t.Errorf("Error adding upstream server: %w", err)
+		t.Errorf("Error adding upstream server: %v", err)
 	}
 	servers, err := c.GetHTTPServers(upstream)
 	if err != nil {
@@ -739,7 +761,7 @@ func TestUpstreamServerDefaultParameters(t *testing.T) {
 	// remove upstream servers
 	_, _, _, err = c.UpdateHTTPServers(upstream, []client.UpstreamServer{})
 	if err != nil {
-		t.Errorf("Couldn't remove servers: %w", err)
+		t.Errorf("Couldn't remove servers: %v", err)
 	}
 }
 
@@ -755,13 +777,13 @@ func TestStreamStats(t *testing.T) {
 	}
 	err = c.AddStreamServer(streamUpstream, server)
 	if err != nil {
-		t.Errorf("Error adding stream upstream server: %w", err)
+		t.Errorf("Error adding stream upstream server: %v", err)
 	}
 
 	// make connection so we have stream server zone stats - ignore response
 	_, err = net.Dial("tcp", helpers.GetStreamAddress())
 	if err != nil {
-		t.Errorf("Error making tcp connection: %w", err)
+		t.Errorf("Error making tcp connection: %v", err)
 	}
 
 	// wait for health checks
@@ -769,7 +791,7 @@ func TestStreamStats(t *testing.T) {
 
 	stats, err := c.GetStats()
 	if err != nil {
-		t.Errorf("Error getting stats: %w", err)
+		t.Errorf("Error getting stats: %v", err)
 	}
 
 	if stats.Connections.Active == 0 {
@@ -806,10 +828,18 @@ func TestStreamStats(t *testing.T) {
 		t.Errorf("Stream upstream 'stream_test' not found")
 	}
 
+	if streamConnLimit, ok := stats.StreamLimitConnections[streamConnZone]; ok {
+		if streamConnLimit.Passed < 1 {
+			t.Errorf("Stream Limit connections stats missing: %v", streamConnLimit.Passed)
+		}
+	} else {
+		t.Errorf("Stream Limit connections %v not found", streamConnLimit)
+	}
+
 	// cleanup stream upstream servers
 	_, _, _, err = c.UpdateStreamServers(streamUpstream, []client.StreamUpstreamServer{})
 	if err != nil {
-		t.Errorf("Couldn't remove stream servers: %w", err)
+		t.Errorf("Couldn't remove stream servers: %v", err)
 	}
 }
 
@@ -838,7 +868,7 @@ func TestStreamUpstreamServerDefaultParameters(t *testing.T) {
 	}
 	err = c.AddStreamServer(streamUpstream, streamServer)
 	if err != nil {
-		t.Errorf("Error adding upstream server: %w", err)
+		t.Errorf("Error adding upstream server: %v", err)
 	}
 	streamServers, err := c.GetStreamServers(streamUpstream)
 	if err != nil {
@@ -857,7 +887,7 @@ func TestStreamUpstreamServerDefaultParameters(t *testing.T) {
 	// cleanup stream upstream servers
 	_, _, _, err = c.UpdateStreamServers(streamUpstream, []client.StreamUpstreamServer{})
 	if err != nil {
-		t.Errorf("Couldn't remove stream servers: %w", err)
+		t.Errorf("Couldn't remove stream servers: %v", err)
 	}
 }
 
@@ -871,7 +901,7 @@ func TestKeyValue(t *testing.T) {
 
 	err = c.AddKeyValPair(zoneName, "key1", "val1")
 	if err != nil {
-		t.Errorf("Couldn't set keyvals: %w", err)
+		t.Errorf("Couldn't set keyvals: %v", err)
 	}
 
 	var keyValPairs client.KeyValPairs
@@ -888,7 +918,7 @@ func TestKeyValue(t *testing.T) {
 
 	keyValuPairsByZone, err := c.GetAllKeyValPairs()
 	if err != nil {
-		t.Errorf("Couldn't get keyvals, %w", err)
+		t.Errorf("Couldn't get keyvals, %v", err)
 	}
 	expectedKeyValPairsByZone := client.KeyValPairsByZone{
 		zoneName: expectedKeyValPairs,
@@ -901,12 +931,12 @@ func TestKeyValue(t *testing.T) {
 	expectedKeyValPairs["key1"] = "valModified1"
 	err = c.ModifyKeyValPair(zoneName, "key1", "valModified1")
 	if err != nil {
-		t.Errorf("couldn't set keyval: %w", err)
+		t.Errorf("couldn't set keyval: %v", err)
 	}
 
 	keyValPairs, err = c.GetKeyValPairs(zoneName)
 	if err != nil {
-		t.Errorf("couldn't get keyval: %w", err)
+		t.Errorf("couldn't get keyval: %v", err)
 	}
 	if !reflect.DeepEqual(expectedKeyValPairs, keyValPairs) {
 		t.Errorf("maps are not equal. expected: %+v, got: %+v", expectedKeyValPairs, keyValPairs)
@@ -920,7 +950,7 @@ func TestKeyValue(t *testing.T) {
 
 	err = c.AddKeyValPair(zoneName, "key2", "val2")
 	if err != nil {
-		t.Errorf("error adding another key/val pair: %w", err)
+		t.Errorf("error adding another key/val pair: %v", err)
 	}
 
 	err = c.DeleteKeyValuePair(zoneName, "key1")
@@ -933,7 +963,7 @@ func TestKeyValue(t *testing.T) {
 	}
 	keyValPairs, err = c.GetKeyValPairs(zoneName)
 	if err != nil {
-		t.Errorf("couldn't get keyval: %w", err)
+		t.Errorf("couldn't get keyval: %v", err)
 	}
 	if !reflect.DeepEqual(keyValPairs, expectedKeyValPairs2) {
 		t.Errorf("didn't delete key1 %+v", keyValPairs)
@@ -941,12 +971,12 @@ func TestKeyValue(t *testing.T) {
 
 	err = c.DeleteKeyValPairs(zoneName)
 	if err != nil {
-		t.Errorf("couldn't delete all: %w", err)
+		t.Errorf("couldn't delete all: %v", err)
 	}
 
 	keyValPairs, err = c.GetKeyValPairs(zoneName)
 	if err != nil {
-		t.Errorf("couldn't get keyval: %w", err)
+		t.Errorf("couldn't get keyval: %v", err)
 	}
 	if len(keyValPairs) > 0 {
 		t.Errorf("zone should be empty after bulk delete")
@@ -969,7 +999,7 @@ func TestKeyValueStream(t *testing.T) {
 
 	err = c.AddStreamKeyValPair(zoneName, "key1", "val1")
 	if err != nil {
-		t.Errorf("Couldn't set keyvals: %w", err)
+		t.Errorf("Couldn't set keyvals: %v", err)
 	}
 
 	keyValPairs, err := c.GetStreamKeyValPairs(zoneName)
@@ -985,7 +1015,7 @@ func TestKeyValueStream(t *testing.T) {
 
 	keyValPairsByZone, err := c.GetAllStreamKeyValPairs()
 	if err != nil {
-		t.Errorf("Couldn't get keyvals, %w", err)
+		t.Errorf("Couldn't get keyvals, %v", err)
 	}
 	expectedKeyValuePairsByZone := client.KeyValPairsByZone{
 		zoneName:       expectedKeyValPairs,
@@ -999,12 +1029,12 @@ func TestKeyValueStream(t *testing.T) {
 	expectedKeyValPairs["key1"] = "valModified1"
 	err = c.ModifyStreamKeyValPair(zoneName, "key1", "valModified1")
 	if err != nil {
-		t.Errorf("couldn't set keyval: %w", err)
+		t.Errorf("couldn't set keyval: %v", err)
 	}
 
 	keyValPairs, err = c.GetStreamKeyValPairs(zoneName)
 	if err != nil {
-		t.Errorf("couldn't get keyval: %w", err)
+		t.Errorf("couldn't get keyval: %v", err)
 	}
 	if !reflect.DeepEqual(expectedKeyValPairs, keyValPairs) {
 		t.Errorf("maps are not equal. expected: %+v, got: %+v", expectedKeyValPairs, keyValPairs)
@@ -1018,7 +1048,7 @@ func TestKeyValueStream(t *testing.T) {
 
 	err = c.AddStreamKeyValPair(zoneName, "key2", "val2")
 	if err != nil {
-		t.Errorf("error adding another key/val pair: %w", err)
+		t.Errorf("error adding another key/val pair: %v", err)
 	}
 
 	err = c.DeleteStreamKeyValuePair(zoneName, "key1")
@@ -1028,7 +1058,7 @@ func TestKeyValueStream(t *testing.T) {
 
 	keyValPairs, err = c.GetStreamKeyValPairs(zoneName)
 	if err != nil {
-		t.Errorf("couldn't get keyval: %w", err)
+		t.Errorf("couldn't get keyval: %v", err)
 	}
 	expectedKeyValPairs2 := client.KeyValPairs{
 		"key2": "val2",
@@ -1039,12 +1069,12 @@ func TestKeyValueStream(t *testing.T) {
 
 	err = c.DeleteStreamKeyValPairs(zoneName)
 	if err != nil {
-		t.Errorf("couldn't delete all: %w", err)
+		t.Errorf("couldn't delete all: %v", err)
 	}
 
 	keyValPairs, err = c.GetStreamKeyValPairs(zoneName)
 	if err != nil {
-		t.Errorf("couldn't get keyval: %w", err)
+		t.Errorf("couldn't get keyval: %v", err)
 	}
 	if len(keyValPairs) > 0 {
 		t.Errorf("zone should be empty after bulk delete")
@@ -1070,7 +1100,7 @@ func TestStreamZoneSync(t *testing.T) {
 
 	err = c1.AddStreamKeyValPair(streamZoneSync, "key1", "val1")
 	if err != nil {
-		t.Errorf("Couldn't set keyvals: %w", err)
+		t.Errorf("Couldn't set keyvals: %v", err)
 	}
 
 	// wait for nodes to sync information of synced zones
@@ -1078,7 +1108,7 @@ func TestStreamZoneSync(t *testing.T) {
 
 	statsC1, err := c1.GetStats()
 	if err != nil {
-		t.Errorf("Error getting stats: %w", err)
+		t.Errorf("Error getting stats: %v", err)
 	}
 
 	if statsC1.StreamZoneSync == nil {
@@ -1118,7 +1148,7 @@ func TestStreamZoneSync(t *testing.T) {
 
 	statsC2, err := c2.GetStats()
 	if err != nil {
-		t.Errorf("Error getting stats: %w", err)
+		t.Errorf("Error getting stats: %v", err)
 	}
 
 	if statsC2.StreamZoneSync == nil {
